Fix for new master
