+++ ftc/mfa/__init__.py
@@ -1,0 +1,1 @@
+from .mfa import MFA
